refactor routes using express Router
